bsrs-django/requirements.txt
Django==1.8.2
django-cors-headers==1.1.0
django-coverage==1.2.4
django-filter==0.10.0
django-nose==1.4.1
djangorestframework==3.1.3
djangorestframework-filters==0.3.0
djangorestframework-jwt==1.6.0
Markdown==2.6.2
model-mommy==1.2.4
nose==1.3.7
psycopg2==2.6.1
PyJWT==1.3.0
six==1.9.0
wheel==0.24.0
